修复胎圈模块数据类型错误问题

+++ Aps-Engine/aps-engine-tq-core/src/main/java/com/zlt/aps/tq/engine/handler/TqBalanceHandler.java
@@ -350,8 +350,8 @@ public class TqBalanceHandler extends AbsTqScheduleStepHandler {
     /**
      * 成型消耗量乘以需求系数
      */
-    private double mulCxPlan(Double cxPlan, double coefficient) {
-        double plan = cxPlan == null ? 0D : cxPlan;
+    private double mulCxPlan(Integer cxPlan, double coefficient) {
+        double plan = cxPlan == null ? 0 : cxPlan;
         return BigDecimalUtil.mul(plan, coefficient);
     }
